WIP - BREAKING - Complete rewrite in progress. See description.
The old way of calling to external executables and tracking devices with a rolling index and a text file was very fragile. I'm moving the get/set device logic into the .Net app and simplifying the flow as much as possible. 
This commit breaks the app in that it still needs implementation for the get/set devices functions and to be tested (especially testing how we callback to reconstruct the menu itself).